+++ backend/backend/models/Cpu.py
@@ -1,0 +1,14 @@
+from pydantic import BaseModel
+
+class Cpu(BaseModel):
+    brand: str
+    model: str
+    fullname: str
+    type: str
+    # Convert ObjectId to string
+    id: str
+
+    class Config:
+        json_encoders = {
+            ObjectId: str  # This will convert ObjectId to a string automatically
+        }
